This is revision 08.12, (pre) Kerniger Kaernter
Outline:
* First, incomplete info.xml version.
* The .deb file is eliminated. Most items are not converted to LOG_DBG() yet
* Up to date SIMP optimization ((multi) frequent piezo with acoustics)
* Ersatz material optimization from Bastian Schmidt
* Compressed textual output 
* VTK export but options to be set in ...SimInOut/VTK/CMakeLists.txt!

Details:
* Removed XML/TestParser
* New schema structure for optimization
* The default saveBegin is now step 0 as the 0th optimization iteration
  is the start configuration
* Schame change: the output formats need no id, it is by default the name
* Added VTK output. It is an option for the ParaView reader plugin
* Transient analysis: you can give a frequency list 
* Now a (incomplete) info.xml file is written. It can be postprocessed
  via XSL (e.g. auxiliary/XSL/extract_eigenvalues_line.xsl):
  xsltproc <stylesheed>.xsl <problem>.info.xml > gnuplot.dat
* The technique is InfoNode, specialization to ParamNode with
  a global "info" pointer. On several occations (precond, solver, ...)
  the current version of the info.xml is written.
* new and changed program options (version, history)
* Ensight is removed
* Assemble::AssembleLinRHS() has no actTime parameter any more as it
  was not used.
* There is no transientharmonic any more
* OLAS has now a InitRHS(Vector<Double>*)

+++ source/OLAS/precond/ILDLPrecond/ildlprecond.cc
@@ -98,7 +98,7 @@ namespace OLAS {
       }
       else {
         if ( this->myParams_->GetBoolValue( "newMatrixPattern" ) == false &&
-             this->sysMatDim_ != scrsMat.GetNcols() ) {
+             (int) this->sysMatDim_ != scrsMat.GetNcols() ) {
           (*error) << "ILDLPrecond::Setup: newMatrixPattern = false, but "
                    << "matrix dimension changed from " << this->sysMatDim_ << " to "
                    << scrsMat.GetNcols();
